Like Button now can Toggle
The like button was being toggled by getting the ElementID "like". Once
Id can only be used once per element and the Like button will be used in
multiple instances, the button must have some sorta of individuality.

The way i found was using the onclick="function(this)" where `this` is
the element that has been clicked. This can then be used inside the
function similar to a `getElementById`.

I then used the toggle inside classList
`this.classList.toggle(className)` that gets the css class in which will
automaticly toggle every time it is called again (Some sort of inside
bool variable).

Cool cool :D